move io utils to their own library, release 1.9.1
