+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: package_config
-version: 1.9.0
+version: 1.9.1
 description: Support for working with Package Configuration files.
 homepage: https://github.com/dart-lang/package_config
 
@@ -14,3 +14,6 @@ dev_dependencies:
   test: ^1.6.4
   matcher: ^0.12.5
   pedantic: 1.8.0
+  build_runner: ^1.0.0
+  build_web_compilers: ^2.0.0
+  build_test: ^0.10.0
