+++ test/discovery_uri_test.dart
@@ -2,6 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+@TestOn('vm')
 library package_config.discovery_test;
 
 import "package:test/test.dart";
